Rear drive semi-fix
due to activaterearclimb drive having true in its "isfinished" command, the command immedietly ends when called which is fine if the command is repeadely called. However, the command is only called once so initialize only runs once. There's better fixes. The current system also disables moving the climblock and movement of the rear leg due to subsystems