button overhaul
now more generic button names that allow for easier remapping later on.  however, various buttons that seem to be subject to change as of currently have been further abstracted to make is so that we don't have to change a ton of specific instances, instead solely changing the one instance that returns the button getting in the single method.
one method to rule each subsystem :)